requirements.txt
cassandra-driver==3.25.0
Flask==1.1.2
webdriver-manager==3.3.0
selenium
requests==2.24.0
gunicorn==19.0.0
